cleanup: rename .H/.CPP to .h/.cpp
